templates/bot-service-adkcode/server/requirements.txt
google-adk[mcp]
google-genai
requests
fastapi
uvicorn
